gate.tsに書き込み
