+++ src/components/mapComponents/MapViewComponent.tsx
@@ -4,7 +4,8 @@ import MapView, { PROVIDER_GOOGLE } from 'react-native-maps';
 import MarkerComponent from './MarkerComponent';
 import PolylineComponent from './PolylineComponent';
 import CustomMap from '../mapComponents/CustomMap';
-import { ToiletMarker, ElevatorMarker, GuideLine, Region, Gate } from '../../domains/map';
+import { ToiletMarker, ElevatorMarker, GuideLine, Region } from '../../domains/map';
+import { Gate } from 'src/domains/gate';
 import { Movie } from '../../domains/movie';
 
 type ScreenNameType = 'video' | 'map';
