+++ src/domains/map.ts
@@ -26,11 +26,3 @@ export interface Region {
   longitudeDelta: number;
 }
 
-export interface Gate {
-  id: number;
-  train_line_name: string;
-  name: string;
-  latitude: number;
-  longitude: number;
-  floor: string;
-}
